+++ elpee/__init__.py
@@ -1,4 +1,4 @@
 from .utils.protocols.st_problem import StandardProblem
 from .utils.protocols.lp_problem import LinearProblem
-from .algorithms.gen_solver import ElpeeSolver
-from .algorithms import alternator as Alternator
+from .elpee_solver import ElpeeSolver as elpee_solver
+from .datahandler import yaml_handler
