+++ elpee/algorithms/__init__.py
@@ -1,14 +1,12 @@
 # elpee/algorithms/__init__.py
 
 from .all_stack_starter import AllStackStarter
-from .alternator import AlternateSolver
 """
 Export Algorithms to solve Linear Programming Problems
 
 Modeules exported by this package:
 
 - `AllStackStarter` : Provide implementation for All Atack Starting Method
-- `AlternateSolver` : Provide implementation for developing alternate solutions
 - `bigM_handler` : Provide implementation for Big M method
 - `DualSimplexSolver` : Provide implementation for Dual Simplex method
